% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/bagEarth2sas.R
\name{bagEarth2sas}
\alias{bagEarth2sas}
\title{Export a bagged MARS model to SAS}
\usage{
bagEarth2sas(fit, name = "prediction", drop = TRUE)
}
\arguments{
\item{fit}{a bagged MARS model trained by \code{\link[caret]{bagEarth}}.}

\item{name}{the name of the variable in which to store the prediction}

\item{drop}{whether to drop the variables for the individual trees}
}
\description{
Generate SAS DATA step code to predict the values of a bagged Multivariate
Adaptive Regression Splines (MARS) model from the \pkg{caret} package.
}
\examples{
require(caret)
trees.bagEarth <- bagEarth(Volume ~ ., data=trees, B=2)
trees.sas <- bagEarth2sas(trees.bagEarth)
cat(trees.sas, file="trees.sas")
}

