% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/earth2sas.R
\name{earth2sas}
\alias{earth2sas}
\title{Export a MARS model to SAS}
\usage{
earth2sas(fit, name = "prediction")
}
\arguments{
\item{fit}{a MARS model trained by \code{\link[earth]{earth}}.  It may
be tuned using \code{\link[caret]{train}}.}

\item{name}{the name of the variable in which to store the prediction}
}
\description{
Generate SAS DATA step code to predict the values of a Multivariate
Adaptive Regression Splines (MARS) model from the \pkg{earth} package.
}
\details{
This function supports only regression with numeric variables, so any
factors must fist be converted to numeric variables (as
\code{\link[caret]{train}} normally does).
}
\examples{
require(earth)
trees.earth <- earth(Volume ~ ., data=trees)
trees.sas <- earth2sas(trees.earth)
cat(trees.sas, file="trees.sas")
}

