% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlmeta.R
\name{mlmeta}
\alias{mlmeta}
\title{Machine Learning Programming for R}
\description{
Export machine learning models trained in R to other programming
languages, so the models can score data in a separate environment
without R.
}

