% Generated by roxygen2 (4.0.2): do not edit by hand
\name{nnet2sas}
\alias{nnet2sas}
\title{Export an artificial neural network to SAS}
\usage{
nnet2sas(fit)
}
\arguments{
\item{fit}{a neural network fitted by nnet()}
}
\description{
Export an artificial neural network to SAS
}
\examples{
require(nnet)
# binary classification
iris2 <- iris[iris$Species \%in\% c('setosa', 'versicolor'),]
iris2$Species <- factor(iris2$Species)
iris2.nnet <- nnet(Species ~ ., data = iris2, size = 1, trace = FALSE)
iris2.sas <- nnet2sas(iris2.nnet)
cat(iris2.sas, file='iris.sas')
}

