% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/simulate_regression_data.R
\name{simulate_regression_data}
\alias{simulate_regression_data}
\title{Generate a Data Set for Testing Regression Models}
\usage{
simulate_regression_data(n = 1000, unordered_factor = TRUE,
  ordered_factor = TRUE, p_missing = 0.03)
}
\arguments{
\item{n}{number of rows}

\item{unordered_factor}{whether to include an unordered factor}

\item{ordered_factor}{whether to include an ordered factor}

\item{p_missing}{the percentage of each predictor to be NA}
}
\value{
a data frame
}
\description{
Generate a Data Set for Testing Regression Models
}
\author{
Greg Ridgeway

Andrew Ziem
}

