% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cforest2sas.R
\name{cforest2sas}
\alias{cforest2sas}
\title{Export a random forest to SAS}
\usage{
cforest2sas(fit, name = "prediction", drop = TRUE)
}
\arguments{
\item{fit}{a random forest trained using \code{\link[party]{cforest}}}

\item{name}{the name of the variable in which to store the prediction}

\item{drop}{whether to drop the variables for the individual trees}
}
\description{
Generate SAS DATA step code to predict the values of a random forest from
the \pkg{party} package.
}
\details{
Unordered factors are supported, while ordered factors and missing values are
not supported.

\code{cforest2sas} averages the predictions of the trees like
\code{\link[randomForest]{randomForest}}, while \code{\link[party]{cforest}}
averages observation weights.
}
\examples{
require(party)
iris.ct <- cforest(Species ~ .,data = iris,
   controls = cforest_unbiased(ntree=5, mtry=2))
iris.sas <- cforest2sas(iris.ct)
cat(iris.sas, file="iris.sas")
}

