% Generated by roxygen2 (4.0.2): do not edit by hand
\name{ctree2sas}
\alias{ctree2sas}
\title{Export a decision tree to SAS}
\usage{
ctree2sas(mytree, node_id = 1, parent_criteria = character(0))
}
\arguments{
\item{mytree}{a decision tree model trained by party::ctree()}

\item{node_id}{the initial node (used internally)}

\item{parent_criteria}{used internally}
}
\description{
Export a decision tree to SAS
}
\examples{
require(party)
iris.ct <- ctree(Species ~ .,data = iris)
iris.sas <- ctree2sas(iris.ct)
cat(iris.sas, file="iris.cat")
}

