% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ctree2sas.R
\name{ctree2sas}
\alias{ctree2sas}
\title{Export a decision tree to SAS}
\usage{
ctree2sas(mytree, name = "prediction", node_id = 1,
  parent_criteria = character(0))
}
\arguments{
\item{mytree}{a decision tree model trained by \code{\link[party]{ctree}}}

\item{name}{the name of the variable in which to store the prediction}

\item{node_id}{the initial node (used internally)}

\item{parent_criteria}{used internally}
}
\description{
Generate SAS DATA step code to predict the values of a conditional inference tree
from the \pkg{party} package.
}
\details{
Missing values and ordinal factors are not supported.
}
\examples{
require(party)
iris.ct <- ctree(Species ~ . , data = iris)
iris.sas <- ctree2sas(iris.ct)
cat(iris.sas, file="iris.sas")
}

