% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/earth2sas.R
\name{earth2sas}
\alias{earth2sas}
\title{Export a MARS model to SAS}
\usage{
earth2sas(fit, name = "prediction", type = "link")
}
\arguments{
\item{fit}{a MARS model trained by \code{\link[earth]{earth}}.  It may
be tuned using \code{\link[caret]{train}}.}

\item{name}{the name of the variable in which to store the prediction}

\item{type}{For classification models \code{type=link} gives log-odds, while
\code{type=response} gives probabilities.}
}
\description{
Generate SAS DATA step code to predict the values of a Multivariate
Adaptive Regression Splines (MARS) model from the \pkg{earth} package.
}
\details{
This function supports regression and binary classification (when
\code{\link[earth]{earth}} is called with \code{glm=list(family=binomial)}.

This function supports only numeric variables, so any factors must first be
converted to numeric variables (as \code{\link[caret]{train}} normally does).

Interactions (\code{degree > 1)} are supported.
}
\examples{
require(earth)
trees.earth <- earth(Volume ~ ., data=trees)
trees.sas <- earth2sas(trees.earth)
cat(trees.sas, file="trees.sas")
}

