% Generated by roxygen2 (4.0.2): do not edit by hand
\name{gbm2sas}
\alias{gbm2sas}
\title{Export a Gradient Boosted Regression Model (GBM) to SAS}
\usage{
gbm2sas(fit, n.trees = fit$n.trees, drop = TRUE)
}
\arguments{
\item{fit}{the model fitted by gbm::gbm()}

\item{n.trees}{the number of trees to export}

\item{drop}{whether to drop the variables for the individual trees}
}
\description{
Export a Gradient Boosted Regression Model (GBM) to SAS
}
\examples{
require(gbm)
mtcars.gbm <- gbm(mpg ~ cyl + disp, data = mtcars, bag.fraction = 1.1,
    distribution = "gaussian")
mtcars.sas <- gbm2sas(mtcars.gbm)
cat(mtcars.sas, file='mtcars.sas')
}

