% Generated by roxygen2 (4.0.2): do not edit by hand
\name{gbm2sas}
\alias{gbm2sas}
\title{Export a Gradient Boosted Regression Model (GBM) to SAS}
\usage{
gbm2sas(fit, n.trees = fit$n.trees)
}
\arguments{
\item{fit}{the model fitted by gbm::gbm()}

\item{n.trees}{the number of trees to export}
}
\description{
Export a Gradient Boosted Regression Model (GBM) to SAS
}

