% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/gbm2sas.R
\name{gbm2sas}
\alias{gbm2sas}
\title{Export a Gradient Boosted Regression Model (GBM) to SAS}
\usage{
gbm2sas(fit, n.trees = fit$n.trees, drop = TRUE)
}
\arguments{
\item{fit}{the model fitted by  \code{\link[gbm]{gbm}}}

\item{n.trees}{the number of trees to export}

\item{drop}{whether to drop the variables for the individual trees}
}
\description{
Generates SAS DATA step code to predict the values of a GBM model.
}
\details{
This function supports unordered factors, ordered factors, and missing
values.  No special considerations are needed for the Gaussian distribution,
and for other distributions, the user is responsible for the link
function in SAS.
}
\examples{
require(gbm)
mtcars.gbm <- gbm(mpg ~ cyl + disp, data = mtcars, bag.fraction = 1.1,
    distribution = "gaussian")
mtcars.sas <- gbm2sas(mtcars.gbm)
cat(mtcars.sas, file='mtcars.sas')
}

