% Generated by roxygen2 (4.0.2): do not edit by hand
\name{simulate_regression_data}
\alias{simulate_regression_data}
\title{Generate a Data Set for Testing Regression Models}
\usage{
simulate_regression_data(n = 1000, unordered_factor = TRUE,
  ordered_factor = TRUE)
}
\arguments{
\item{n}{number of rows}

\item{unordered_factor}{whether to include an unordered factor}

\item{ordered_factor}{whether to include an ordered factor}
}
\value{
a data frame
}
\description{
Generate a Data Set for Testing Regression Models
}
\author{
Greg Ridgeway

Andrew Ziem
}

